requirements.txt
# BASE
boto3==1.24.24
django-environ==0.9.0
django-storages==1.12.3
Django==4.0.6
djangorestframework==3.13.1
Pillow==9.2.0
psycopg2-binary==2.9.3
python-dotenv==0.20.0
requests==2.28.1

# PROD
gunicorn==20.1.0

# FRONTEND
# https://github.com/materializecss/materialize v1.1.0
